Develop

+++ src/main/java/ru/yandex/practicum/filmorate/storage/dao/MpaStorageDao.java
@@ -7,7 +7,7 @@ import java.util.Optional;
 
 public interface MpaStorageDao {
 
-    Optional<Mpa> findById(int id);
+    Optional<Mpa> findById(long id);
 
     List<Mpa> findAll();
 }
