+++ src/main/java/ru/yandex/practicum/filmorate/storage/impl/GenreStorageDaoImpl.java
@@ -11,7 +11,6 @@ import java.util.*;
 
 @Repository
 public class GenreStorageDaoImpl implements GenreStorageDao {
-
     private final JdbcTemplate jdbcTemplate;
 
     public GenreStorageDaoImpl(JdbcTemplate jdbcTemplate) {
@@ -49,9 +48,8 @@ public class GenreStorageDaoImpl implements GenreStorageDao {
 
     private Genre makeLocalGenre(ResultSet rs) throws SQLException {
         return new Genre(
-                rs.getInt("GENRE_ID"),
+                rs.getLong("GENRE_ID"),
                 rs.getString("GENRE_NAME")
         );
     }
-
 }
