+++ src/main/java/ru/yandex/practicum/filmorate/storage/impl/MpaStorageDaoImpl.java
@@ -12,7 +12,6 @@ import java.util.Optional;
 
 @Repository
 public class MpaStorageDaoImpl implements MpaStorageDao {
-
     private final JdbcTemplate jdbcTemplate;
 
     public MpaStorageDaoImpl(JdbcTemplate jdbcTemplate) {
@@ -20,7 +19,7 @@ public class MpaStorageDaoImpl implements MpaStorageDao {
     }
 
     @Override
-    public Optional<Mpa> findById(int id) {
+    public Optional<Mpa> findById(long id) {
         final String sqlQuery = "SELECT * FROM MPA WHERE MPA_ID = ?";
         final List<Mpa> mpa = jdbcTemplate.query(sqlQuery, this::makeLocalMpa, id);
         return mpa.size() == 0 ?
@@ -36,7 +35,7 @@ public class MpaStorageDaoImpl implements MpaStorageDao {
 
     private Mpa makeLocalMpa(ResultSet rs, int rowNum) throws SQLException {
         return new Mpa(
-                rs.getInt("MPA_ID"),
+                rs.getLong("MPA_ID"),
                 rs.getString("MPA_NAME")
         );
     }
